+++ GardeningHoe.h
@@ -17,7 +17,7 @@ class GardeningHoe: public Equipment{
         void get_Function();
         void Upgrade();
         int get_util_lvl();
-        double get_boost_val();
+        int get_boost_val();
         string get_name();
 
 };
